updated responsibilities
